feat: renamed to db_utilities
